tests: add tests

+++ src/StevesBot.Worker/Discord/DiscordClientOptions.cs
@@ -4,5 +4,5 @@ internal sealed class DiscordClientOptions
 {
   public string ApiUrl { get; init; } = string.Empty;
   public string AppToken { get; init; } = string.Empty;
-  public int Intents { get; init; }
+  public long Intents { get; init; }
 }
